+++ components/src/core/drawer/drawer-header.tsx
@@ -9,6 +9,7 @@ import {
     ImageStyle,
     TextStyle,
     ViewProps,
+    PixelRatio,
 } from 'react-native';
 import { H6, Subtitle1 } from '../typography';
 import { Divider, useTheme } from 'react-native-paper';
@@ -16,26 +17,38 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { EdgeInsets } from '../__types__';
 import { $DeepPartial } from '@callstack/react-theme-provider';
 
-const makeStyles = (props: DrawerHeaderProps, theme: ReactNativePaper.Theme, insets: EdgeInsets): any =>
-    StyleSheet.create({
+const makeStyles = (
+    props: DrawerHeaderProps,
+    theme: ReactNativePaper.Theme,
+    insets: EdgeInsets
+): StyleSheet.NamedStyles<{
+    root: ViewStyle;
+    icon: ViewStyle;
+    content: ViewStyle;
+    textContent: ViewStyle;
+    title: TextStyle;
+    subtitle: TextStyle;
+    backgroundImageWrapper: ViewStyle;
+    backgroundImage: ImageStyle;
+}> => {
+    const fontScale = PixelRatio.getFontScale();
+    return StyleSheet.create({
         root: {
             paddingTop: insets.top,
             // @ts-ignore
             backgroundColor: props.backgroundColor || theme.colors.primaryBase || theme.colors.primary,
         },
         icon: {
-            height: 56,
-            width: 52,
-            paddingLeft: 4,
+            height: 56 * fontScale,
+            width: 24 * fontScale + 32,
             justifyContent: 'center',
         },
         content: {
             flexDirection: 'row',
         },
         textContent: {
-            color: 'red',
             flexDirection: 'column',
-            padding: 4,
+            paddingVertical: 4 * fontScale,
             paddingLeft: 16,
             flex: 1,
             height: '100%',
@@ -47,7 +60,7 @@ const makeStyles = (props: DrawerHeaderProps, theme: ReactNativePaper.Theme, ins
         subtitle: {
             color: props.fontColor || theme.colors.surface,
             lineHeight: 16,
-            marginTop: -2,
+            marginTop: -2 * fontScale,
         },
         backgroundImageWrapper: {
             position: 'absolute',
@@ -63,6 +76,7 @@ const makeStyles = (props: DrawerHeaderProps, theme: ReactNativePaper.Theme, ins
             resizeMode: 'cover',
         },
     });
+};
 
 export type DrawerHeaderProps = ViewProps & {
     /** Colored background of the header */
@@ -140,6 +154,7 @@ export const DrawerHeader: React.FC<DrawerHeaderProps> = (props) => {
                     <Image
                         source={backgroundImage}
                         resizeMethod={'resize'}
+                        // @ts-ignore typescript is being weird about the backgroundImage style type
                         style={[defaultStyles.backgroundImage, styles.backgroundImage]}
                     />
                 </View>